+++ task/worker.hpp
@@ -5,7 +5,7 @@ class Worker : public UniversityPerson
 {
     size_t money_;
     public:
-    int GetMoney() override {return money_;};
+    size_t GetMoney() override {return money_;};
     void SetMoney(size_t money){money_ = money;} 
 
     Worker(std::string name, std::string surname, std::string adr, int id, Gender gender, size_t money);
